Replace incremental execution updates by briefings
Sending incremental updates on execution progress to the web client
didn't work if client didn't view the Run page the entire time the
execution was running or when the page was open in multiple browser
windows/tabs.

We now send the entire execution log and status as 'briefing' on client
request making communication between client and server more robust.