Fix adding relationships
A reference to recently added relationship's objects was passed around
Entity editor to EntityDiff where it was stored to be later committed
to the model database. However, the same list could be modified by
adding new relationships to the model. This led to a situation where
the Entity editor erroneously though that user was trying to add new
relationships with the same objects. We now clone the object list in
an appropriate place to get rid of the troublesome references.

Re #9