Fix changing relationship objects
Entity editor was complaining about clashing relationship objects when
a relationship was modified such that it actually clashed with another
one. It may seem sensible behavior but in fact ofter prevented the
modification of relationships. We now allow relationships to be in a
clashing state temporarily and prevent commits while in the state.

Re #9